LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_setpointercapture_pointerup_mouse-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Events Capture Test - setPointerCapture on pointerup

Test Description: This test checks if setPointerCapture works properly. Complete the following actions:
Press and hold left mouse button over black rectangle
Release left mouse button

Test passes if the proper behavior of the events is observed.

Pointer Events Capture Test

Test complete: Scroll to Summary to view Pass/Fail Results.

The following pointer types were detected: (none).

The following events were logged: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented
